fix exception caused by save of nls tabs in separate code window
